requirements.txt
discord.py
supabase
google-generativeai
pypdf
python-dotenv
